[refactor] #100 Participant 도메인 리팩토링
